fix: routes and email modals
