+++ app/pages/catalogo/DescartavelPage.js
@@ -1,9 +1,9 @@
 import React, { useState, useEffect } from 'react';
 import { View, Text, StyleSheet, ScrollView, ActivityIndicator, ImageBackground, Dimensions } from 'react-native';
 import { useFonts, Poppins_400Regular, Poppins_500Medium, Poppins_600SemiBold } from '@expo-google-fonts/poppins';
-import Header from '../components/Header';
-import Footer from '../components/Footer';
-import Banner from '../assets/backgrounds/bannerEspecial.png'; 
+import Header from '../../components/Header';
+import Footer from '../../components/Footer';
+import Banner from '../../assets/backgrounds/bannerEspecial.png'; 
 
 const { width } = Dimensions.get('window');
