+++ app/routes/AppNavigator.js
@@ -2,17 +2,17 @@ import React from 'react';
 import { createBottomTabNavigator } from '@react-navigation/bottom-tabs';
 import Home from '../pages/Home';
 import Trilha from '../pages/Trilha';
-import Catalogo from '../pages/Catalogo';
+import Catalogo from '../pages/catalogo/Catalogo';
 import Coleta from '../pages/Coleta';
 import Perfil from '../pages/Perfil';
-import DescartavelPage from '../pages/DescartavelPage'
-import ReciclavelPage from '../pages/ReciclavelPage'
-import Login from '../pages/Login';
-import NoticiasPage from '../pages/NoticiasPage';
+import DescartavelPage from '../pages/catalogo/DescartavelPage'
+import ReciclavelPage from '../pages/catalogo/ReciclavelPage'
+import Login from '../pages/login/Login';
+import NoticiasPage from '../pages/catalogo/NoticiasPage';
 import Config from '../pages/sidebar/Config';
-import RedefinirSenha from '../pages/RedefinirSenha';
-import Token from '../pages/Token';
-import NovaSenha from '../pages/NovaSenha';
+import RedefinirSenha from '../pages/login/RedefinirSenha';
+import Token from '../pages/login/Token';
+import NovaSenha from '../pages/login/NovaSenha';
 
 // Cria o Tab Navigator
 const Tab = createBottomTabNavigator();
